// src/store/DictionaryContext/enums.ts
export enum EDictionaryEntityId {
  AssetPool = 'AssetPool',
  BalanceSheet = 'BalanceSheet',
  Calendar = 'Calendar',
  Classification = 'Classification',
  Contract = 'Contract',
  Contributor = 'Contributor',
  DataSource = 'DataSource',
  Document = 'Document',
  Employer = 'Employer',
  Engagement = 'Engagement',
  ExcellentEmployer = 'ExcellentEmployer',
  Expectation = 'Expectation',
  Fee = 'Fee',
  FinancialStatement_Postings = 'FinancialStatement_Postings',
  Individual = 'Individual',
  Ledger_Valuations = 'Ledger_Valuations',
  LegalEntity = 'LegalEntity',
  LegalEntityRole = 'LegalEntityRole',
  Login = 'Login',
  MarketClassification = 'MarketClassification',
  Organisation = 'Organisation',
  ParticipantLedger_Position = 'ParticipantLedger_Position',
  ParticipantLedger_Trade = 'ParticipantLedger_Trade',
  Person = 'Person',
  PersonRole = 'PersonRole',
  PostingRule = 'PostingRule',
  References = 'References',
  Relationship = 'Relationship',
  Self = 'Self',
  ValuationPolicy = 'ValuationPolicy',
}

export enum EDictionaryEntryType {
  Action = 'action',
  Attribute = 'attribute',
  Decorator = 'decorator',
}